pyproject.toml
[project]
name = "explainer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jupyter>=1.1.1",
    "lab>=8.5",
    "openai>=2.8.0",
    "playwright>=1.56.0",
    "strands-agents[anthropic]>=1.16.0",
    "strands-agents-tools[use-browser,user-browser]>=0.2.15",
    "python-dotenv>=1.2.1",
    "chromadb>=1.3.7",
    "langchain-anthropic>=1.3.0",
    "langgraph>=1.0.5",
    "lxml>=6.0.2",
    "pytest-playwright>=0.7.2",
    "langchain>=1.2.0",
    "browser-use>=0.11.2",
    "langchain-community>=0.4.1",
]
